+++ xsl_erp_dao/src/main/java/com/xsl/erp/mapper/XslDatetimeMapper.java
@@ -6,7 +6,7 @@ import xsl.erp.pojo.XslDatetime;
 import xsl.erp.pojo.XslDatetimeExample;
 
 public interface XslDatetimeMapper {
-    int countByExample(XslDatetimeExample example);
+    long countByExample(XslDatetimeExample example);
 
     int deleteByExample(XslDatetimeExample example);
