+++ xsl_erp_dao/src/main/java/com/xsl/erp/mapper/XslMatchRewardMapper.java
@@ -6,7 +6,7 @@ import xsl.erp.pojo.XslMatchReward;
 import xsl.erp.pojo.XslMatchRewardExample;
 
 public interface XslMatchRewardMapper {
-    int countByExample(XslMatchRewardExample example);
+    long countByExample(XslMatchRewardExample example);
 
     int deleteByExample(XslMatchRewardExample example);
