+++ xsl_erp_dao/src/main/java/com/xsl/erp/mapper/XslSchoolMapper.java
@@ -6,7 +6,7 @@ import xsl.erp.pojo.XslSchool;
 import xsl.erp.pojo.XslSchoolExample;
 
 public interface XslSchoolMapper {
-    int countByExample(XslSchoolExample example);
+    long countByExample(XslSchoolExample example);
 
     int deleteByExample(XslSchoolExample example);
